requirements.txt
pandas
sqlalchemy
pyodbc
elasticsearch
PyYAML
cx_Oracle